make pytest-bdd and pytest-benchmark defer imports too
